functions/data_leads/requirements.txt
attrs==19.3.0
boto3==1.9.219
botocore==1.12.219
certifi==2019.6.16
chardet==3.0.4
coverage==5.2
docutils==0.14
Faker==2.0.0
idna==2.8
importlib-metadata==1.6.1
ipaddress==1.0.22
jmespath==0.9.4
kafka-python==1.4.7
lxml==4.4.0
more-itertools==8.3.0
numpy==1.18.5
packaging==20.4
pandas==0.24.2
pluggy==0.13.1
psycopg2==2.7.6.1
py==1.8.1
pycrypto==2.6.1
pydash==4.7.5
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
python-dateutil==2.8.0
python-dotenv==0.14.0
python-snappy==0.5.4
pytz==2019.2
PyYAML==5.2
requests==2.22.0
requests-mock==1.5.2
s3transfer==0.2.1
six==1.12.0
text-unidecode==1.2
urllib3==1.25.3
wcwidth==0.2.4
xmltodict==0.12.0
zipp==3.1.0
